Extracted InlineLogs component
